% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmw-package.R
\docType{package}
\name{lmw-package}
\alias{lmw-package}
\alias{_PACKAGE}
\title{lmw: Linear Model Weights}
\description{
Computes the implied weights of linear regression models for estimating average causal effects and provides diagnostics based on these weights. These diagnostics rely on the analyses in Chattopadhyay and Zubizarreta (2022) \doi{10.1093/biomet/asac058} where several regression estimators are represented as weighting estimators, in connection to inverse probability weighting. 'lmw' provides tools to diagnose representativeness, balance, extrapolation, and influence for these models, clarifying the target population of inference. Tools are also available to simplify estimating treatment effects for specific target populations of interest.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ngreifer/lmw}
  \item Report bugs at \url{https://github.com/ngreifer/lmw/issues}
}

}
\author{
\strong{Maintainer}: Noah Greifer \email{ngreifer@iq.harvard.edu} (\href{https://orcid.org/0000-0003-3067-7154}{ORCID})

Authors:
\itemize{
  \item Ambarish Chattopadhyay \email{ambarish_chattopadhyay@g.harvard.edu} (\href{https://orcid.org/0000-0002-1502-0974}{ORCID})
  \item Jose Zubizarreta \email{zubizarreta@hcp.med.harvard.edu} (\href{https://orcid.org/0000-0002-0322-147X}{ORCID})
}

}
\keyword{internal}
